Move all token cache implementations into .token_cache
Previously, implementations lived in their platform specific files. However, this created an import cycle that pylint was complaining about.
