+++ tests/test_windows_backend.py
@@ -10,7 +10,8 @@ import msal
 if not sys.platform.startswith('win'):
     pytest.skip('skipping windows-only tests', allow_module_level=True)
 else:
-    from msal_extensions.windows import WindowsDataProtectionAgent, WindowsTokenCache
+    from msal_extensions.windows import WindowsDataProtectionAgent
+    from msal_extensions.token_cache import WindowsTokenCache
 
 
 def test_dpapi_roundtrip_with_entropy():
